Implement the WinHttpGetDefaultProxyConfiguration function for the Windows proxy search strategy.
This function checks for proxy settings at the registry (WinHTTP) level, as opposed to the IE search strategy, which checks at the user (WinINet) level. The proxy isn't always set at the WinHTTP level, so Bernd Rosstauscher's fallback was to use the IE search strategy for both, but now both functions are available.