Fix SOCKS proxy settings not being returned (closes #10)
Although SOCKS proxy settings are detected correctly, the
ProtocolDispatchSelector tries to lookup a selector using the protocol
as the key, which doesn't work for SOCKS because "socks://" isn't part
of the URL.

The fix is to specify it as a fallback instead, so that it is
automatically used in the absence of a more protocol specific (e.g.
http) setting.