Added support for DHCP resolution to WpadProxySearchStrategy
Made a couple minor changes to the DHCPSocket and DHCPOptions classes to
make them able to handle Option 252 and capable of sending on a specific
interface. Then used them to send DHCPInform messages on all interfaces
with valid IPv4 addresses and check the responses for Option 252. Also
added WPAD to the Strategy enum so it can be tested with ProxyTester and
added as a Strategy.